+++ Sandswept/Equipment/EquipmentBase.cs
@@ -112,8 +112,49 @@ namespace Sandswept.Equipment
             LanguageAPI.Add("EQUIPMENT_SANDSWEPT_" + EquipmentLangTokenName + "_DESCRIPTION", EquipmentFullDescription);
             LanguageAPI.Add("EQUIPMENT_SANDSWEPT_" + EquipmentLangTokenName + "_LORE", EquipmentLore);
 
+            if (EquipmentModel != null)
+            {
+                CreateModelPanelParameters(EquipmentModel);
+            }
+
             ItemAPI.Add(new CustomEquipment(EquipmentDef, CreateItemDisplayRules()));
             On.RoR2.EquipmentSlot.PerformEquipmentAction += PerformEquipmentAction;
+        }
+
+        private void CreateModelPanelParameters(GameObject equipmentModel)
+        {
+            if (equipmentModel.GetComponent<ModelPanelParameters>() != null)
+            {
+                return;
+            }
+
+            GameObject model = PrefabAPI.InstantiateClone(equipmentModel, equipmentModel.name + "-fixed", false);
+            GameObject focus = new("Focus");
+            GameObject camera = new("Camera");
+            MeshRenderer biggestRenderer = model.GetComponentsInChildren<MeshRenderer>().ToList().OrderByDescending(x => ToFloat(x.bounds.size)).First();
+            float mult = ToFloat(biggestRenderer.bounds.size) / 3f;
+            float min = mult;
+            float max = 3f * mult;
+            focus.transform.parent = model.transform;
+            camera.transform.parent = model.transform;
+            focus.transform.position = biggestRenderer.bounds.center;
+            camera.transform.localPosition = focus.transform.position + (model.transform.forward * max);
+
+            var modelPanelParameters = model.AddComponent<ModelPanelParameters>();
+            modelPanelParameters.focusPointTransform = focus.transform;
+            modelPanelParameters.cameraPositionTransform = camera.transform;
+            modelPanelParameters.minDistance = min;
+            modelPanelParameters.maxDistance = max;
+
+            EquipmentDef.pickupModelPrefab = model;
+        }
+
+        public static float ToFloat(Vector3 vec)
+        {
+            vec.x = Mathf.Abs(vec.x);
+            vec.y = Mathf.Abs(vec.y);
+            vec.z = Mathf.Abs(vec.z);
+            return vec.x + vec.y + vec.z;
         }
 
         private bool PerformEquipmentAction(On.RoR2.EquipmentSlot.orig_PerformEquipmentAction orig, RoR2.EquipmentSlot self, EquipmentDef equipmentDef)